+++ src/__tests__/EThree.test.ts
@@ -5,7 +5,6 @@ import {
     CardManager,
     VirgilCardVerifier,
     GeneratorJwtProvider,
-    IKeyEntry,
     CachingJwtProvider,
 } from 'virgil-sdk';
 import {
@@ -90,7 +89,7 @@ describe('VirgilE2ee', () => {
     const fetchToken = () => Promise.resolve(generator.generateToken(identity).toString());
 
     it('full integration test', async done => {
-        const sdk = await EThree.init(fetchToken);
+        const sdk = await EThree.initialize(fetchToken);
         const password = 'secret_password';
         const cloudStorage = await createSyncStorage(identity, password);
         await sdk.bootstrap();
@@ -126,7 +125,7 @@ describe('local bootstrap (without password)', () => {
     it('AUTH-1 has no local key, has no card', async done => {
         const identity = 'virgiltestlocal1' + Date.now();
         const fetchToken = createFetchToken(identity);
-        const sdk = await EThree.init(fetchToken);
+        const sdk = await EThree.initialize(fetchToken);
         await sdk.bootstrap();
         const cards = await cardManager.searchCards(identity);
         expect(cards.length).toEqual(1);
@@ -136,7 +135,7 @@ describe('local bootstrap (without password)', () => {
     it('has local key, has no card', async done => {
         const identity = 'virgiltestlocal2' + Date.now();
         const fetchToken = createFetchToken(identity);
-        const sdk = await EThree.init(fetchToken);
+        const sdk = await EThree.initialize(fetchToken);
         await sdk.bootstrap();
         const cards = await cardManager.searchCards(identity);
         expect(cards.length).toEqual(1);
@@ -156,7 +155,7 @@ describe('local bootstrap (without password)', () => {
 
         expect(prevCards.length).toEqual(1);
 
-        const sdk = await EThree.init(fetchToken);
+        const sdk = await EThree.initialize(fetchToken);
         await sdk.bootstrap();
         const cards = await cardManager.searchCards(identity);
 
@@ -170,7 +169,7 @@ describe('local bootstrap (without password)', () => {
         const keyPair = virgilCrypto.generateKeys();
         await cardManager.publishCard({ identity: identity, ...keyPair });
         const fetchToken = createFetchToken(identity);
-        const sdk = await EThree.init(fetchToken);
+        const sdk = await EThree.initialize(fetchToken);
         const cards = await cardManager.searchCards(identity);
 
         expect(cards.length).toEqual(1);
@@ -189,7 +188,7 @@ describe('local bootstrap (without password)', () => {
         const fetchToken = createFetchToken(identity);
         const prevCards = await cardManager.searchCards(identity);
         expect(prevCards.length).toBe(0);
-        const sdk = await EThree.init(fetchToken);
+        const sdk = await EThree.initialize(fetchToken);
         await sdk.bootstrap();
         const [cards, key] = await Promise.all([
             cardManager.searchCards(identity),
@@ -209,7 +208,7 @@ describe('remote bootstrap (with password)', () => {
 
         expect(prevCards.length).toBe(0);
 
-        const sdk = await EThree.init(fetchToken);
+        const sdk = await EThree.initialize(fetchToken);
         await sdk.bootstrap('secure_password');
         const [cards, key] = await Promise.all([
             cardManager.searchCards(identity),
@@ -229,7 +228,7 @@ describe('remote bootstrap (with password)', () => {
         const cloudStorage = await createSyncStorage(identity, 'secret_password');
         await cloudStorage.storeEntry(identity, virgilCrypto.exportPrivateKey(keyPair.privateKey));
         expect(prevCards.length).toBe(1);
-        const sdk = await EThree.init(fetchToken);
+        const sdk = await EThree.initialize(fetchToken);
         await sdk.bootstrap('secure_password');
         const [cards, key] = await Promise.all([
             cardManager.searchCards(identity),
@@ -254,7 +253,7 @@ describe('remote bootstrap (with password)', () => {
         await keyknoxStorage.remove(identity);
         await keyStorage.remove(identity);
         expect(prevCards.length).toBe(1);
-        const sdk = await EThree.init(fetchToken);
+        const sdk = await EThree.initialize(fetchToken);
         try {
             await sdk.bootstrap('not_secret_password');
         } catch (e) {
@@ -269,8 +268,23 @@ describe('lookupKeys', () => {
     const identity = 'virgiltestlookup' + Date.now();
     const fetchToken = () => Promise.resolve(generator.generateToken(identity).toString());
 
+    it('lookupKeys for one identity success', async done => {
+        const sdk = await EThree.initialize(fetchToken);
+        const identity = 'virgiltestlookup' + Date.now();
+        const keypair = virgilCrypto.generateKeys();
+
+        await cardManager.publishCard({ identity: identity, ...keypair });
+        const publicKey = await sdk.lookupPublicKeys(identity);
+
+        expect(Array.isArray(publicKey)).not.toBeTruthy();
+        expect(virgilCrypto.exportPublicKey(publicKey).toString('base64')).toEqual(
+            virgilCrypto.exportPublicKey(keypair.publicKey).toString('base64'),
+        );
+        done();
+    });
+
     it('STE-1 lookupKeys success', async done => {
-        const sdk = await EThree.init(fetchToken);
+        const sdk = await EThree.initialize(fetchToken);
         const identity1 = 'virgiltestlookup1' + Date.now();
         const identity2 = 'virgiltestlookup2' + Date.now();
         const identity3 = 'virgiltestlookup3' + Date.now();
@@ -299,7 +313,7 @@ describe('lookupKeys', () => {
     });
 
     it('lookupKeys nonexistent identity', async done => {
-        const sdk = await EThree.init(fetchToken);
+        const sdk = await EThree.initialize(fetchToken);
         const identity1 = 'virgiltestlookupnonexist' + Date.now();
         const identity2 = 'virgiltestlookupnonexist' + Date.now();
         try {
@@ -347,7 +361,7 @@ describe('lookupKeys', () => {
     });
 
     it('STE-2 lookupKeys with empty array of identities', async done => {
-        const sdk = await EThree.init(fetchToken);
+        const sdk = await EThree.initialize(fetchToken);
         try {
             await sdk.lookupPublicKeys([]);
         } catch (e) {
@@ -363,7 +377,7 @@ describe('change password', () => {
         const identity = 'virgiltest' + Date.now();
         const fetchToken = () => Promise.resolve(generator.generateToken(identity).toString());
         try {
-            const sdk = await EThree.init(fetchToken);
+            const sdk = await EThree.initialize(fetchToken);
             await sdk.changePassword('old_password', 'new_password');
             await sdk.cleanup();
             await sdk.bootstrap('new_password');
@@ -378,7 +392,7 @@ describe('change password', () => {
         const identity = 'virgiltest' + Date.now();
         const fetchToken = () => Promise.resolve(generator.generateToken(identity).toString());
         try {
-            const sdk = await EThree.init(fetchToken);
+            const sdk = await EThree.initialize(fetchToken);
             await sdk.bootstrap('old_password');
             await sdk.changePassword('old_password', 'new_password');
             await sdk.cleanup();
@@ -396,7 +410,7 @@ describe('backupPrivateKey', () => {
     const fetchToken = () => Promise.resolve(generator.generateToken(identity).toString());
 
     it('success', async done => {
-        const sdk = await EThree.init(fetchToken);
+        const sdk = await EThree.initialize(fetchToken);
         const storage = await createSyncStorage(identity, 'secret_password');
 
         try {
@@ -417,7 +431,7 @@ describe('backupPrivateKey', () => {
     });
 
     it('fail', async done => {
-        const sdk = await EThree.init(fetchToken);
+        const sdk = await EThree.initialize(fetchToken);
         try {
             await sdk.bootstrap('secret_pass');
             await sdk.backupPrivateKey('secret_pass');
@@ -441,11 +455,9 @@ describe('encrypt and decrypt', () => {
         const fetchToken2 = () => Promise.resolve(generator.generateToken(identity2).toString());
 
         const [sdk1, sdk2] = await Promise.all([
-            EThree.init(fetchToken1),
-            EThree.init(fetchToken2),
+            EThree.initialize(fetchToken1),
+            EThree.initialize(fetchToken2),
         ]);
-
-        const unusedKeypair = virgilCrypto.generateKeys();
 
         await Promise.all([sdk1.bootstrap(), sdk2.bootstrap()]);
         const message = 'encrypt, decrypt, repeat';
@@ -453,17 +465,17 @@ describe('encrypt and decrypt', () => {
         const sdk2PublicKeys = await sdk2.lookupPublicKeys([identity2]);
         const encryptedMessage = await sdk1.encrypt(message, sdk2PublicKeys);
         try {
-            await sdk2.decrypt(encryptedMessage, [unusedKeypair.publicKey]);
+            await sdk2.decrypt(encryptedMessage);
         } catch (e) {
             expect(e).toBeInstanceOf(Error);
         }
-        const decryptedMessage = await sdk2.decrypt(encryptedMessage, sdk1PublicKeys);
+        const decryptedMessage = await sdk2.decrypt(encryptedMessage, sdk1PublicKeys[0]);
         expect(decryptedMessage).toEqual(message);
         done();
     });
 
     it('STE-4 encrypt for empty public keys', async done => {
-        const sdk = await EThree.init(fetchToken);
+        const sdk = await EThree.initialize(fetchToken);
         await sdk.bootstrap();
         try {
             await sdk.encrypt('privet', []);
@@ -474,22 +486,8 @@ describe('encrypt and decrypt', () => {
         done('should throw');
     });
 
-    it('STE-5 decrypt for empty public keys', async done => {
-        const sdk = await EThree.init(fetchToken);
-        await sdk.bootstrap();
-        const keyPair = virgilCrypto.generateKeys();
-        const message = await sdk.encrypt('privet', [keyPair.publicKey]);
-        try {
-            await sdk.decrypt(message, []);
-        } catch (e) {
-            expect(e).toBeInstanceOf(EmptyArrayError);
-            return done();
-        }
-        done('should throw');
-    });
-
     it('STE-6 encrypt and decrypt without public keys', async done => {
-        const sdk = await EThree.init(fetchToken);
+        const sdk = await EThree.initialize(fetchToken);
         await sdk.bootstrap();
         const message = 'secret message';
         const encryptedMessage = await sdk.encrypt(message);
@@ -499,7 +497,7 @@ describe('encrypt and decrypt', () => {
     });
 
     it('STE-7 decrypt message without sign', async done => {
-        const sdk = await EThree.init(fetchToken);
+        const sdk = await EThree.initialize(fetchToken);
         await sdk.bootstrap();
         const receiverPublicKey = await sdk.lookupPublicKeys([identity]);
         const { publicKey: senderPublicKey } = virgilCrypto.generateKeys();
@@ -508,7 +506,7 @@ describe('encrypt and decrypt', () => {
             .encrypt(message, receiverPublicKey)
             .toString('base64');
         try {
-            await sdk.decrypt(encryptedMessage, [senderPublicKey]);
+            await sdk.decrypt(encryptedMessage, senderPublicKey);
         } catch (e) {
             expect(e).toBeDefined();
             return done();
@@ -518,7 +516,7 @@ describe('encrypt and decrypt', () => {
 
     it('STE-8 no decrypt/encrypt before bootstrap', async done => {
         await keyStorage.clear();
-        const sdk = await EThree.init(fetchToken);
+        const sdk = await EThree.initialize(fetchToken);
         try {
             await sdk.encrypt('message');
         } catch (e) {
@@ -539,13 +537,13 @@ describe('encrypt and decrypt', () => {
         const buf = new Buffer('123');
 
         const recipient = virgilCrypto.generateKeys();
-        const sdk = await EThree.init(fetchToken);
+        const sdk = await EThree.initialize(fetchToken);
         await sdk.bootstrap();
-        const publicKeys = await sdk.lookupPublicKeys([identity]);
+        const publicKey = (await sdk.lookupPublicKeys([identity]))[0];
         const encryptedMessage = await sdk.encrypt(buf, [recipient.publicKey]);
         expect(encryptedMessage).toBeInstanceOf(Buffer);
 
-        const resp = await sdk.decrypt(encryptedMessage, publicKeys);
+        const resp = await sdk.decrypt(encryptedMessage, publicKey);
         expect(resp).toBeInstanceOf(Buffer);
         done();
     });
@@ -556,7 +554,7 @@ describe('cleanup()', () => {
         const identity = 'virgiltestlogout' + Date.now();
         const fetchToken = () => Promise.resolve(generator.generateToken(identity).toString());
 
-        const sdk = await EThree.init(fetchToken);
+        const sdk = await EThree.initialize(fetchToken);
         await sdk.bootstrap('secure_password');
         const isDeleted = await sdk.cleanup();
         const privateKeyFromLocalStorage = await keyStorage.load(identity);
@@ -571,7 +569,7 @@ describe('cleanup()', () => {
         const identity = 'virgiltestlogout' + Date.now();
         const fetchToken = () => Promise.resolve(generator.generateToken(identity).toString());
 
-        const sdk = await EThree.init(fetchToken);
+        const sdk = await EThree.initialize(fetchToken);
         await sdk.bootstrap('secure_password');
         await sdk.resetPrivateKeyBackup('secure_password');
         try {